Move the <form> tags into InputSettings and RadioSettings
