+++ frontend/form-builder/src/components/FormSettings/ButtonGroup.tsx
@@ -1,5 +1,5 @@
 import React, { useContext } from "react";
-import classes from "./ComponentSettings.module.css";
+import classes from "./FormSettings.module.css";
 import { Button } from "@digdir/design-system-react";
 import { FloppydiskFillIcon, TrashFillIcon } from "@navikt/aksel-icons";
 import { FormBuilderContext } from "../../context";
