+++ frontend/form-builder/src/components/FormSettings/InputSettings.tsx
@@ -1,55 +1,80 @@
 import { useTranslation } from "react-i18next";
-import React, { useContext, useState } from "react";
+import React, { FormEvent, useContext, useState } from "react";
 import { Checkbox, Radio, Textfield } from "@digdir/design-system-react";
 import { ButtonGroup } from "./ButtonGroup";
 import { FormBuilderContext } from "../../context";
+import classes from "./FormSettings.module.css";
+import { cleanFormData } from "../../../../main-app/src/components/FormViewer/validationUtils";
+import { ComponentAsStrings } from "./types";
+import { updateComponent, updateComponentArray } from "./utils";
 
 export const InputSettings = () => {
   const { t } = useTranslation();
-  const { selectedItem } = useContext(FormBuilderContext);
+  const { selectedItem, setSelectedItem, form, setForm } = useContext(FormBuilderContext);
   const [inputType, setInputType] = useState(selectedItem.inputType || "string");
 
   const handleInputFormat = (value: string) => {
     setInputType(value as "string" | "number");
   };
 
+  const handleSaveComponent = (event: FormEvent<HTMLFormElement>) => {
+    event.preventDefault();
+
+    const formData = new FormData(event.currentTarget as HTMLFormElement);
+    const updatedComponentData = cleanFormData(formData) as ComponentAsStrings;
+
+    // TODO: Use Zod validation here.
+    // See FormViewer for an example of how it can be done.
+
+    const updatedComponent: FormComponent = updateComponent(selectedItem, updatedComponentData);
+    const updatedComponents: FormComponent[] = updateComponentArray(
+      form.components,
+      updatedComponent,
+    );
+
+    setForm({ ...form, components: updatedComponents });
+    setSelectedItem(null);
+  };
+
   return (
     <>
-      <Textfield
-        name="name"
-        label={t("settings_side_bar.component.name")}
-        defaultValue={selectedItem.name || ""}
-        size="small"
-        placeholder={"Required"}
-      />
-      <Textfield
-        name="label"
-        label={t("settings_side_bar.component.label")}
-        defaultValue={selectedItem.label || ""}
-        size="small"
-        placeholder={"Required"}
-      />
-      <Checkbox
-        name="required"
-        size="small"
-        value="required"
-        defaultChecked={selectedItem.required}
-      >
-        {t("settings_side_bar.require.response")}
-      </Checkbox>
-      <Radio.Group
-        name="inputType"
-        defaultValue={selectedItem.inputType || inputType}
-        size="small"
-        legend={t("settings_side_bar.input.format")}
-        onChange={handleInputFormat}
-      >
-        <Radio value="string">{t("settings_side_bar.text")}</Radio>
-        <Radio value="number">{t("settings_side_bar.number")}</Radio>
-      </Radio.Group>
-      {inputType === "string" && <TextSettings />}
-      {inputType === "number" && <NumberSettings />}
-      <ButtonGroup />
+      <form className={classes.SettingsContent} onSubmit={handleSaveComponent}>
+        <Textfield
+          name="name"
+          label={t("settings_side_bar.component.name")}
+          defaultValue={selectedItem.name || ""}
+          size="small"
+          placeholder={"Required"}
+        />
+        <Textfield
+          name="label"
+          label={t("settings_side_bar.component.label")}
+          defaultValue={selectedItem.label || ""}
+          size="small"
+          placeholder={"Required"}
+        />
+        <Checkbox
+          name="required"
+          size="small"
+          value="required"
+          defaultChecked={selectedItem.required}
+        >
+          {t("settings_side_bar.require.response")}
+        </Checkbox>
+        <Radio.Group
+          name="inputType"
+          defaultValue={selectedItem.inputType || inputType}
+          size="small"
+          legend={t("settings_side_bar.input.format")}
+          onChange={handleInputFormat}
+        >
+          <Radio value="string">{t("settings_side_bar.text")}</Radio>
+          <Radio value="number">{t("settings_side_bar.number")}</Radio>
+        </Radio.Group>
+        {inputType === "string" && <TextSettings />}
+        {inputType === "number" && <NumberSettings />}
+        <ButtonGroup />
+      </form>
     </>
   );
 };
