+++ frontend/form-builder/src/components/FormSettings/SettingsSidebar.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { Heading } from "@digdir/design-system-react";
 import { useTranslation } from "react-i18next";
-import classes from "./ComponentSettings.module.css";
+import classes from "./FormSettings.module.css";
 
 interface SettingsSidebarProps {
   SettingsContent: () => React.JSX.Element;
